requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via
    #   -r requirements.in
    #   django
build==1.2.1
    # via
    #   -r requirements.in
    #   pip-tools
cfgv==3.4.0
    # via
    #   -r requirements.in
    #   pre-commit
click==8.1.7
    # via
    #   -r requirements.in
    #   pip-tools
distlib==0.3.8
    # via
    #   -r requirements.in
    #   virtualenv
django==5.0.7
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.15.2
    # via -r requirements.in
filelock==3.15.4
    # via
    #   -r requirements.in
    #   virtualenv
identify==2.6.0
    # via
    #   -r requirements.in
    #   pre-commit
nodeenv==1.9.1
    # via
    #   -r requirements.in
    #   pre-commit
packaging==24.1
    # via
    #   -r requirements.in
    #   build
pillow==10.4.0
    # via -r requirements.in
pip-tools==7.4.1
    # via -r requirements.in
platformdirs==4.2.2
    # via
    #   -r requirements.in
    #   virtualenv
pre-commit==3.8.0
    # via -r requirements.in
psycopg2==2.9.9
    # via -r requirements.in
pyparsing==3.1.2
    # via -r requirements.in
pyproject-hooks==1.1.0
    # via
    #   -r requirements.in
    #   build
    #   pip-tools
pyyaml==6.0.1
    # via
    #   -r requirements.in
    #   pre-commit
sqlparse==0.5.1
    # via
    #   -r requirements.in
    #   django
virtualenv==20.26.3
    # via
    #   -r requirements.in
    #   pre-commit
wheel==0.43.0
    # via
    #   -r requirements.in
    #   pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
